app_backend/requirements.txt
attrs==21.4.0
black==22.3.0
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
qrcode==7.3.1
tomli==2.0.1
typing-extensions==4.1.1
Werkzeug==2.1.1
zipp==3.8.0